Move "async" utils to own module.
