+++ src/client/common/core.utils.ts
@@ -3,9 +3,5 @@
 
 'use strict';
 
-export async function sleep(timeout: number) {
-    return new Promise(resolve => setTimeout(resolve, timeout));
-}
-
 // tslint:disable-next-line:no-empty
 export function noop() { }
