+++ src/client/common/utils/async.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+'use strict';
+
+export async function sleep(timeout: number) {
+    return new Promise((resolve) => {
+        setTimeout(resolve, timeout);
+    });
+}
