+++ src/test/debugger/capabilities.test.ts
@@ -14,8 +14,8 @@ import { PassThrough } from 'stream';
 import { Message } from 'vscode-debugadapter/lib/messages';
 import { DebugProtocol } from 'vscode-debugprotocol';
 import { EXTENSION_ROOT_DIR } from '../../client/common/constants';
-import { sleep } from '../../client/common/core.utils';
 import { createDeferred } from '../../client/common/helpers';
+import { sleep } from '../../client/common/utils/async';
 import { PTVSD_PATH } from '../../client/debugger/Common/constants';
 import { ProtocolParser } from '../../client/debugger/Common/protocolParser';
 import { ProtocolMessageWriter } from '../../client/debugger/Common/protocolWriter';
